rename some files since they went from component to hook
